added a Delete button in the contacts list
added a delete button in the contacts list main display. when this delete button is clicked, the main contactslist controller calls the Contacts Factory deletecontact method and when the delete is successfuly, the child view is loaded to display the message received from the server. Also the contacts list in contactslist controller is updated to show that the contact has been removed.